introduce-request-log-manager(move): Hand request dumping over to RequestLogManager
Pure physical move per MECH_COMMIT_SPLIT. Cut @staticmethod dump_requests
+ record_request_for_crash_dump + _dump_data_to_file +
dump_requests_before_crash from TokenizerManager; paste into
RequestLogManager (drop @staticmethod, replace
``self: "RequestLogManager"`` → plain ``self``). Discard the now-dead
init_request_logging_and_dumping (factory does the work). Caller prefix
replacement: ``TokenizerManager.<method>(self.request_log_manager, ...)``
→ ``self.request_log_manager.<method>(...)``.

Refactor chain ID: introduce-request-log-manager-move
PR-Title: Extract RequestLogManager from TokenizerManager
